se agrega clv a managment y operaciones (#3)
* se agrega clv a managment y operaciones

* se borra variable h
